# openapi.yaml
...
openapi: 3.0.3
info:
  title: rucQ API
  version: 0.1.0
  description: rucQ API
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/events/{event_id}/register:
    post:
      summary: イベントに参加
      tags:
        - Events
      operationId: registerEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: イベントへの参加をキャンセル
      tags:
        - Events
      operationId: unregisterEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/events/{event_id}/participants:
    get:
      summary: イベントの参加者一覧を取得
      tags:
        - Events
      operationId: getParticipants
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/question_groups:
    get:
      summary: 質問グループの一覧を取得
      tags:
        - Questions
      operationId: getQuestionGroups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionGroup"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: 質問グループを作成
      tags:
        - Questions
      operationId: postQuestionGroup
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostQuestionGroupRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/question_groups/{question_group_id}:
    get:
      summary: 質問グループの詳細を取得
      tags:
        - Questions
      operationId: getQuestionGroup
      parameters:
        - $ref: "#/components/parameters/QuestionGroupId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionGroup"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: 質問グループを更新
      tags:
        - Questions
      operationId: putQuestionGroup
      parameters:
        - $ref: "#/components/parameters/QuestionGroupId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostQuestionGroupRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: 質問グループを削除
      tags:
        - Questions
      operationId: deleteQuestionGroup
      parameters:
        - $ref: "#/components/parameters/QuestionGroupId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/questions:
    get:
      summary: 質問の一覧を取得
      tags:
        - Questions
      operationId: getQuestions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: 質問を作成
      tags:
        - Questions
      operationId: postQuestion
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostQuestionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/questions/{question_id}:
    get:
      summary: 質問の詳細を取得
      tags:
        - Questions
      operationId: getQuestion
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: 質問を更新
      tags:
        - Questions
      operationId: putQuestion
      parameters:
        - $ref: "#/components/parameters/QuestionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostQuestionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: 質問を削除
      tags:
        - Questions
      operationId: deleteQuestion
      parameters:
        - $ref: "#/components/parameters/QuestionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    X-Forwarded-User:
      name: X-Forwarded-User
      in: header
      description: ログインしているユーザーのtraQ ID（NeoShowcaseが自動で付与）
      schema:
        type: string
...
    EventId:
      name: event_id
      in: path
      description: イベントID
      required: true
      schema:
        type: integer
    QuestionGroupId:
      name: question_group_id
      in: path
      description: 質問グループのID
      required: true
      schema:
        type: integer
    QuestionId:
      name: question_id
      in: path
      description: 質問ID
      required: true
      schema:
        type: integer
...
  responses:
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
  schemas:
...
    User:
      type: object
      properties:
        traq_id:
          type: string
        is_staff:
          type: boolean
      required:
        - traq_id
        - is_staff
    QuestionGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        due:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        camp_id:
          type: integer
      required:
        - id
        - name
        - description
        - due
        - questions
        - camp_id
    PostQuestionGroupRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        due:
          type: string
          format: date-time
        camp_id:
          type: integer
      required:
        - name
        - description
        - due
        - camp_id
    Question:
      type: object
      properties:
        id:
          type: integer
        question_group_id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum:
            - single
            - multiple
            - free_text
            - free_number
        is_public:
          type: boolean
        is_open:
          type: boolean
        options:
          type: array
          items:
            $ref: "#/components/schemas/Option"
          nullable: true
      required:
        - id
        - question_group_id
        - title
        - description
        - type
        - is_public
        - is_open
    PostQuestionRequest:
      type: object
      properties:
        question_group_id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
            - single
            - multiple
            - free_text
            - free_number
        is_public:
          type: boolean
        is_open:
          type: boolean
      required:
        - question_group_id
        - title
        - description
        - type
        - is_public
        - is_open
    Option:
      type: object
      properties:
        id:
          type: integer
        question_id:
          type: integer
        content:
          type: string
      required:
        - id
        - question_id
        - content
...